[#4] refactor: Change package structure for Entity
1. domain.entity에 entity 클래스가 모여있습니다.
2. domain.entity.aggregate.[aggregate root entity 이름] 으로 aggregate에 속한 entity를 모아놨습니다.
3. domain.entity.lookup 에 lookup 테이블과 관련된 entity를 모아놨습니다.
