+++ BE/sidedish/src/main/java/team16/sidedish/repository/ProviderRepository.java
@@ -1,7 +1,7 @@
 package team16.sidedish.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import team16.sidedish.domain.entity.provider.Provider;
+import team16.sidedish.domain.entity.aggregate.provider.Provider;
 
 public interface ProviderRepository extends CrudRepository<Provider, Long> {
 }
